requirements.txt
asgiref==3.8.1
attrs==23.2.0
colorama==0.4.6
Django==4.2.13
django-bootstrap5==22.2
djangorestframework==3.15.2
flake8==7.0.0
flake8-isort==6.1.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.0
pep8-naming==0.13.3
pluggy==1.4.0
psycopg2==2.9.9
py==1.11.0
pycodestyle==2.11.1
pyflakes==3.2.0
pygame==2.5.2
pytest==7.1.3
pytz==2024.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.1
tzdata==2024.1
django-debug-toolbar=4.4.5